refactor(StringSeparator) : 문자열 구분기 패키지 이동
