--- src/main/java/me/devyonghee/calculator/StringsProvider.java
@@ -1,8 +1,0 @@
-package me.devyonghee.calculator;
-
-import java.util.List;
-
-interface StringsProvider {
-
-    List<String> strings();
-}
