+++ src/main/java/me/devyonghee/common/StringsProvider.java
@@ -1,0 +1,8 @@
+package me.devyonghee.common;
+
+import java.util.List;
+
+public interface StringsProvider {
+
+    List<String> strings();
+}
